Stablize game sockets, removing certain queries
