+++ client/app/battle/battle.js
@@ -1,12 +1,14 @@
 angular.module('battlescript.battle', [])
 
-.controller('BattleController', function($scope, $timeout, Battle) {
+.controller('BattleController', function($scope, $timeout, Battle, SocketHolder) {
+
+  console.log('hello');
   
   ////////////////////////////////////////////////////////////
   // init players
   ////////////////////////////////////////////////////////////
 
-  $scope.playerOne = window.localStorage.getItem('username');
+  $scope.playerOne = SocketHolder.playerOne;
   $scope.playerTwo = "...";
 
 
@@ -17,9 +19,8 @@ angular.module('battlescript.battle', [])
   // open up socket and handle socket events
   ////////////////////////////////////////////////////////////
 
-  var socket = io('http://localhost:8000', {
-    query: 'username=' + $scope.playerOne + '&handler=battle'
-  });
+  var socket = SocketHolder.socket;
+  socket.emit('matchReady');
 
   socket.emit('updateUsers');
   socket.on('userList', function(userArray){
