+++ server/server.js
@@ -22,15 +22,12 @@ server.listen(8000);
 var io = require('socket.io')(server);
 
 // set up two handlers for separate sockets
-var battleHandler = require('./config/battleHandler.js');
-var dashboardHandler = require('./config/dashboardHandler.js');
+var socketHandler = require('./config/socketHandler.js');
 
 
 // For handling various sockets, goto socket battleHandler in config js
 io.on('connection', function(socket){
-  var handler = socket.handshake.query.handler;
-  if (handler === 'battle') battleHandler(socket, io);
-  if (handler === 'dashboard') dashboardHandler(socket, io);
+  socketHandler(socket, io);
 });
 
 // export our app for testing and flexibility, required by index.js
